legal-etl/pyproject.toml
[tool.poetry]
name = "legal-etl"
version = "0.1.0"
description = "Legal document ingest, versioning, embedding, and indexing pipeline."
authors = ["Legal ETL Team <dev@legal-etl.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.2"
tenacity = "^9.0.0"
selectolax = "^0.3.20"
lxml = "^5.3.0"
pydantic = "^2.8.2"
pymupdf = "^1.24.10"
"pdfminer.six" = "^20240706"
python-dateutil = "^2.9.0.post0"
orjson = "^3.10.7"
psycopg = {version = "^3.2.1", extras = ["binary"]}
qdrant-client = "^1.11.1"
opensearch-py = "^2.7.1"
redis = "^5.0.7"
rq = "^1.16.2"
tqdm = "^4.66.4"
structlog = "^24.4.0"
cohere = "^5.6.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"
ruff = "^0.5.4"

[tool.poetry.group.playwright.dependencies]
playwright = "^1.46.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true
